+++ x/erc20/proposal_handler.go
@@ -64,14 +64,13 @@ func handleRegisterCoinProposal(
 			return err
 		}
 
-		err = ctx.EventManager().EmitTypedEvent(&types.EventRegisterPair{
-			Denom:        pair.Denom,
-			Erc20Address: pair.Erc20Address,
-		})
-
-		if err != nil {
-			k.Logger(ctx).Error(err.Error())
-		}
+		ctx.EventManager().EmitEvent(
+			sdk.NewEvent(
+				types.EventTypeRegisterCoin,
+				sdk.NewAttribute(types.AttributeKeyCosmosCoin, pair.Denom),
+				sdk.NewAttribute(types.AttributeKeyERC20Token, pair.Erc20Address),
+			),
+		)
 	}
 
 	return nil
@@ -90,14 +89,13 @@ func handleRegisterERC20Proposal(
 			return err
 		}
 
-		err = ctx.EventManager().EmitTypedEvent(&types.EventRegisterPair{
-			Denom:        pair.Denom,
-			Erc20Address: pair.Erc20Address,
-		})
-
-		if err != nil {
-			k.Logger(ctx).Error(err.Error())
-		}
+		ctx.EventManager().EmitEvent(
+			sdk.NewEvent(
+				types.EventTypeRegisterERC20,
+				sdk.NewAttribute(types.AttributeKeyCosmosCoin, pair.Denom),
+				sdk.NewAttribute(types.AttributeKeyERC20Token, pair.Erc20Address),
+			),
+		)
 	}
 
 	return nil
@@ -114,14 +112,13 @@ func handleToggleConversionProposal(
 		return err
 	}
 
-	err = ctx.EventManager().EmitTypedEvent(&types.EventToggleTokenConversion{
-		Denom:        pair.Denom,
-		Erc20Address: pair.Erc20Address,
-	})
-
-	if err != nil {
-		k.Logger(ctx).Error(err.Error())
-	}
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			types.EventTypeToggleTokenConversion,
+			sdk.NewAttribute(types.AttributeKeyCosmosCoin, pair.Denom),
+			sdk.NewAttribute(types.AttributeKeyERC20Token, pair.Erc20Address),
+		),
+	)
 
 	return nil
 }