+++ x/ibc/transfer/ibc_module.go
@@ -17,8 +17,8 @@
 package transfer
 
 import (
-	ibctransfer "github.com/cosmos/ibc-go/v5/modules/apps/transfer"
-	porttypes "github.com/cosmos/ibc-go/v5/modules/core/05-port/types"
+	ibctransfer "github.com/cosmos/ibc-go/v6/modules/apps/transfer"
+	porttypes "github.com/cosmos/ibc-go/v6/modules/core/05-port/types"
 	"github.com/evmos/evmos/v11/x/ibc/transfer/keeper"
 )