+++ x/recovery/keeper/utils_test.go
@@ -1,6 +1,8 @@
 package keeper_test
 
 import (
+	"context"
+
 	"github.com/stretchr/testify/mock"
 
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
@@ -8,8 +10,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 
-	transfertypes "github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
-	clienttypes "github.com/cosmos/ibc-go/v5/modules/core/02-client/types"
+	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
 
 	"github.com/evmos/evmos/v11/x/recovery/types"
 )
@@ -30,22 +31,21 @@ func (m *MockTransferKeeper) GetDenomTrace(ctx sdk.Context, denomTraceHash tmbyt
 	return args.Get(0).(transfertypes.DenomTrace), args.Bool(1)
 }
 
-func (m *MockTransferKeeper) SendTransfer(
-	ctx sdk.Context,
-	sourcePort,
-	sourceChannel string,
-	token sdk.Coin,
-	sender sdk.AccAddress,
-	receiver string,
-	timeoutHeight clienttypes.Height,
-	timeoutTimestamp uint64,
-) error {
-	args := m.Called(mock.Anything, sourcePort, sourceChannel, token, mock.Anything, mock.Anything, mock.Anything, mock.Anything)
-
-	err := m.SendCoinsFromAccountToModule(ctx, sender, transfertypes.ModuleName, sdk.Coins{token})
+func (m *MockTransferKeeper) Transfer(goCtx context.Context, msg *transfertypes.MsgTransfer) (*transfertypes.MsgTransferResponse, error) {
+	args := m.Called(
+		mock.Anything,
+		mock.Anything,
+	)
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	sender, err := sdk.AccAddressFromBech32(msg.Sender)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	return args.Error(0)
+	err = m.SendCoinsFromAccountToModule(ctx, sender, transfertypes.ModuleName, sdk.Coins{msg.Token})
+	if err != nil {
+		return nil, err
+	}
+
+	return nil, args.Error(1)
 }